refactor: create shared lists
